Decorations: minor refactoring, no logical changes intended
