+++ include/decorations.h
@@ -1,0 +1,9 @@
+/* SPDX-License-Identifier: GPL-2.0-only */
+#ifndef __LAB_DECORATIONS_H
+#define __LAB_DECORATIONS_H
+
+struct server;
+
+void xdg_server_decoration_init(struct server *server);
+
+#endif /* __LAB_DECORATIONS_H */
